requirements.txt
APScheduler==2.1.2
asn1crypto==0.24.0
cached-property==1.3.0
cffi==1.12.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
DBUtils==1.3
defusedxml==0.5.0
Django==1.10
django-api-doc==0.4
django-api-docs==1.1.1
django-cors-headers==2.4.0
django-filter==2.0.0.dev1
django-pipeline==1.6.13
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
futures==3.1.1
ipaddress==1.0.17
IPy==0.83
isodate==0.5.4
itypes==1.1.0
Jinja2==2.10.1
ldap3==2.6
lxml==3.8.0
Markdown==3.1
MarkupSafe==1.1.1
ntlm-auth==1.3.0
pyasn1==0.1.9
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
pymssql==2.1.4
PyMySQL==0.7.9
pyotp==2.2.7
pypsrp==0.3.1
pytz==2017.2
requests==2.12.4
requests-toolbelt==0.8.0
setuptools==41.0.1
six==1.10.0
uritemplate==3.0.0
uWSGI==2.0.13.1
wakeonlan==0.2.2
xlrd==1.0.0
XlsxWriter==1.1.8
xlwt==1.3.0
zeep==1.2.0
tzlocal==1.5.1